Pull down remote configuration
When this script is executed it attempts to go to the URL in the
configuration file and pull down a configuration word which specifies
which GPIO pins should be set to high and low, then sets the pins. It
will only allow a pull-down if the device requesting it has an
authentication token; otherwise it will post an access denied.